requirements.txt
beautifulsoup4==4.9.1
Django==2.2.1
django-bootstrap4==1.1.0
django-stdimage==5.1.1
importlib-metadata==1.7.0
Pillow==7.2.0
progressbar2==3.53.1
python-utils==2.4.0
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0
flake8==3.8.3